added background to header menu, made refactoring
