+++ src/components/shared/BuyMeCoffeeLink.tsx
@@ -12,7 +12,7 @@ export const BuyMeCoffeeLink = ({ className }: BuyMeCoffeeLinkProps) => {
             href={BUY_ME_COFFEE}
             target="_blank"
             rel="noopener noreferrer"
-            className={`pointer relative w-[56px] h-[56px] flex justify-center items-center
+            className={`pointer relative w-12 h-12 pc:w-[56px] pc:h-[56px] flex justify-center items-center
             after:content-[''] after:absolute after:w-[0px] after:h-[0px] after:rounded-full after:bg-purple-stroke03 
             after:scale-0 after:origin-center after:blur-[2px] after:transition-scale after:ease-out after:duration-[600ms]
             hover:after:scale-100 hover:after:blur-[2px] hover:after:w-full hover:after:h-full
